build 20200713

+++ src/libxpeccy/filetypes/t64.c
@@ -120,7 +120,7 @@ int loadC64RawTap(Computer* comp, const char* name, int dsk) {
 						per = fgett(file);
 						// printf("%i\n",per);
 					}
-					per = (int)(8 * per / comp->cpuFrq);		// comp->cpuFrq - MHz (1e6), result period in mks (1e-6)
+					per = 8 * per / comp->cpuFrq;		// comp->cpuFrq - MHz (1e6), result period in mks (1e-6)
 					if (per > 1e5) {
 						blkAddPause(&blk, per);		// pause
 						tapAddBlock(comp->tape, blk);
